move installer into a service

+++ tests/TestCase/Command/InstallCommandTest.php
@@ -4,6 +4,10 @@ namespace SwaggerBake\Test\TestCase\Command;
 
 use Cake\TestSuite\ConsoleIntegrationTestTrait;
 use Cake\TestSuite\TestCase;
+use SwaggerBake\Lib\Exception\InstallException;
+use SwaggerBake\Lib\Exception\SwaggerBakeRunTimeException;
+use SwaggerBake\Lib\Service\InstallerService;
+use SwaggerBake\Lib\Service\OpenApiBakerService;
 
 class InstallCommandTest extends TestCase
 {
@@ -21,16 +25,24 @@ class InstallCommandTest extends TestCase
         $this->useCommandRunner();
 
         $this->configDir = CONFIG . 'testing' . DS;
-        foreach (scandir($this->configDir) as $file) {
+        $files = scandir($this->configDir);
+        if (!is_array($files)) {
+            throw new \RuntimeException("Tests cannot be run because no files were found.");
+        }
+        foreach ($files as $file) {
             if (is_file($this->configDir . $file)) {
                 unlink($this->configDir . $file);
             }
         }
     }
 
-    public function test(): void
+    public function test_install_should_work(): void
     {
-        $this->exec('swagger install --config_test ' . $this->configDir, ['Y','/']);
+        $this->mockService(InstallerService::class, function () {
+            return new InstallerService($this->configDir);
+        });
+
+        $this->exec('swagger install', ['Y', '/']);
         $this->assertOutputContains('Installation Complete!');
         $this->assertFileExists($this->configDir . 'swagger.yml');
         $this->assertFileExists($this->configDir . 'swagger_bake.php');
@@ -42,29 +54,42 @@ class InstallCommandTest extends TestCase
         $this->assertOutputContains('Exiting install');
     }
 
-    public function test_do_not_overwrite_existing(): void
+    public function test_user_can_skip_error(): void
     {
-        $this->exec('swagger install', ['Y','N']);
-        $this->assertOutputContains('SwaggerBake Install');
-        $this->assertExitError();
+        $this->mockService(InstallerService::class, function () {
+            $mock = $this->createMock(InstallerService::class);
+            $matcher = $this->exactly(2);
+            $mock
+                ->expects($matcher)
+                ->method('install')
+                ->withAnyParameters()
+                ->willReturnCallback(function () use ($matcher) {
+                    if ($matcher->getInvocationCount() === 1) {
+                        throw (new InstallException())->setQuestion("skip me");
+                    }
+
+                    return true;
+                });
+            return $mock;
+        });
+
+        $this->exec('swagger install', ['Y', '/', 'Y']);
+        $this->assertOutputContains('Installation Complete!');
     }
 
-    public function test_assets_directory_not_found(): void
+    public function test_exception_is_printed_as_console_error(): void
     {
-        $this->exec('swagger install --config_test /config/no-exists --assets_test /nope');
-        $this->assertErrorContains('Unable to locate assets directory, please install manually');
-        $this->assertExitError();
-    }
+        $this->mockService(InstallerService::class, function () {
+            $mock = $this->createMock(InstallerService::class);
+            $mock
+                ->expects($this->once())
+                ->method('install')
+                ->withAnyParameters()
+                ->willThrowException(new InstallException("error message"));
+            return $mock;
+        });
 
-    public function test_file_copy_fails(): void
-    {
-        $this->exec('swagger install --config_test /config/no-exists', ['Y','/']);
-        $this->assertExitError();
-    }
-
-    public function test_invalid_api_path_prefix(): void
-    {
-        $this->exec('swagger install --config_test ' . $this->configDir, ['Y','\ sdf sdf ']);
-        $this->assertExitError();
+        $this->exec('swagger install', ['Y', '/']);
+        $this->assertOutputContains('error message');
     }
 }
